fix for reported bug by DC-1: crash in station gui
the template gui item was added to the drop-down list that was also
shown in a station gui, but there was no action present when this item
was selected in a station gui. per default the game would commit suicide
by called NOT_REACHED() at the default case of the according
switch-statement.