Fixed crashes on auto- and template-replace
Fixed a segfault that occured when a train that was scheduled to be both
auto-renewed and template-replaced entered a depot. This happened due to
the auto-renew, which occured first, invalidating the pointer to the
train that was later used by template replacement. There was a sanity
check in place for this, but critical access took place before the
sanity check. Fixed by moving critical access to the invalid pointer
behind the sanity check.