Followup fix: Queues for template auto- and template replacement
If a train is scheduled for both auto- and template-replacement, then
the intent of the user should always be to template-replace the train,
but in case that fails for whatever reason the train should still be
auto-replaced. This commit makes sure that a train is still
auto-replaced if template-replacement fails -- which is supposed to be
when it doesn't cost anything.
THIS MAY OR MAY NOT STILL SEGFAULT depending on whether auto-replacement
does sufficient sanity checks. The situation in which it might segfault
is when we try to auto-replace a train that was incorrectly detected to
not be template replaced, which may happen if it was template replaced
with costs of 0 (i.e. all needed engines / waggons were already in the
depot and none had to be bought)