reordered includes
1. system includes
2. general openttd includes
3. project includes

The 3 lists are ordered alphabetically as far as possible without leading to
any compilation errors. The window* includes were creating problems and have
been moved to the top of the includes lists.

fixed compilation due to header interdependencies

With the window_type.h located in the engine_gui.h instead of .cpp, we are able
to sort the includes in tbtr.

This must be some of the code copied directly from engine_gui.cpp.

sorted includes in tbtr_template_gui_replaceall.h

sorted includes in 2 .cpp files

sorted includes in tbtr_template_vehicle_func.cpp

reverted include ordering in engine_gui.*

This makes it necessary to move the window* includes further up in some source
files.