fixed crash during template replacement
The reason is decribed in a comment contained in this commit.

Since this change is related to the test-buy for a template train before
the actual replacement is performed, a notification about the
incompleteness of this function (TestBuyAll...) has been added. This
will need to be extended in the future.